backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
elasticsearch
kafka-python
clickhouse-driver 